fix: price 변수 전부 Int 로 변경

+++ src/main/java/com/example/hanaro/domain/order/entity/Order.java
@@ -41,7 +41,7 @@ public class Order extends BaseEntity {
     private OrderStatus status;
 
     @Column(name = "total_price", nullable = false)
-    private BigDecimal totalPrice;
+    private int totalPrice;
 
 
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
